# openapi/openapi.yaml
openapi: 3.0.0
paths:
...
  /v1/kyc/receivers:
    post:
      operationId: ReceiversController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReceiverDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiverEntity"
        "400":
          description: Validation failed, or the request is not valid in the current state.
          content:
            application/json:
              schema: *a1
        "401":
          description: No authenticated consumer, or admin credentials are required.
          content:
            application/json:
              schema: *a1
        "403":
          description: The API key lacks the required scope, or the request did not arrive
            through the gateway.
          content:
            application/json:
              schema: *a1
        "404":
          description: The resource does not exist, or does not belong to this consumer.
          content:
            application/json:
              schema: *a1
        "409":
          description: Conflicts with existing state — an idempotency clash, an operation
            already in flight, or an invalid state transition.
          content:
            application/json:
              schema: *a1
        "500":
          description: Unexpected server error. The detail is logged server-side and never
            returned.
          content:
            application/json:
              schema: *a1
        "502":
          description: An upstream provider returned an error.
          content:
            application/json:
              schema: *a1
        "503":
          description: An upstream provider is unavailable, or the service is misconfigured.
          content:
            application/json:
              schema: *a1
      summary: Create a receiver (start KYC/KYB)
      tags: &a11
        - kyc
    get:
      operationId: ReceiversController_findAll_v1
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiverListEntity"
        "400":
          description: Validation failed, or the request is not valid in the current state.
          content:
            application/json:
              schema: *a1
        "401":
          description: No authenticated consumer, or admin credentials are required.
          content:
            application/json:
              schema: *a1
        "403":
          description: The API key lacks the required scope, or the request did not arrive
            through the gateway.
          content:
            application/json:
              schema: *a1
        "404":
          description: The resource does not exist, or does not belong to this consumer.
          content:
            application/json:
              schema: *a1
        "409":
          description: Conflicts with existing state — an idempotency clash, an operation
            already in flight, or an invalid state transition.
          content:
            application/json:
              schema: *a1
        "500":
          description: Unexpected server error. The detail is logged server-side and never
            returned.
          content:
            application/json:
              schema: *a1
        "502":
          description: An upstream provider returned an error.
          content:
            application/json:
              schema: *a1
        "503":
          description: An upstream provider is unavailable, or the service is misconfigured.
          content:
            application/json:
              schema: *a1
      summary: List the consumer's receivers
      tags: *a11
  /v1/kyc/receivers/{id}:
    get:
      operationId: ReceiversController_findOne_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiverEntity"
        "400":
          description: Validation failed, or the request is not valid in the current state.
          content:
            application/json:
              schema: *a1
        "401":
          description: No authenticated consumer, or admin credentials are required.
          content:
            application/json:
              schema: *a1
        "403":
          description: The API key lacks the required scope, or the request did not arrive
            through the gateway.
          content:
            application/json:
              schema: *a1
        "404":
          description: The resource does not exist, or does not belong to this consumer.
          content:
            application/json:
              schema: *a1
        "409":
          description: Conflicts with existing state — an idempotency clash, an operation
            already in flight, or an invalid state transition.
          content:
            application/json:
              schema: *a1
        "500":
          description: Unexpected server error. The detail is logged server-side and never
            returned.
          content:
            application/json:
              schema: *a1
        "502":
          description: An upstream provider returned an error.
          content:
            application/json:
              schema: *a1
        "503":
          description: An upstream provider is unavailable, or the service is misconfigured.
          content:
            application/json:
              schema: *a1
      summary: Get a receiver (refreshes KYC status from BlindPay)
      tags: *a11
    patch:
      operationId: ReceiversController_update_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReceiverDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiverEntity"
        "400":
          description: Validation failed, or the request is not valid in the current state.
          content:
            application/json:
              schema: *a1
        "401":
          description: No authenticated consumer, or admin credentials are required.
          content:
            application/json:
              schema: *a1
        "403":
          description: The API key lacks the required scope, or the request did not arrive
            through the gateway.
          content:
            application/json:
              schema: *a1
        "404":
          description: The resource does not exist, or does not belong to this consumer.
          content:
            application/json:
              schema: *a1
        "409":
          description: Conflicts with existing state — an idempotency clash, an operation
            already in flight, or an invalid state transition.
          content:
            application/json:
              schema: *a1
        "500":
          description: Unexpected server error. The detail is logged server-side and never
            returned.
          content:
            application/json:
              schema: *a1
        "502":
          description: An upstream provider returned an error.
          content:
            application/json:
              schema: *a1
        "503":
          description: An upstream provider is unavailable, or the service is misconfigured.
          content:
            application/json:
              schema: *a1
      summary: Update a receiver
      tags: *a11
    delete:
      operationId: ReceiversController_remove_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
        "400":
          description: Validation failed, or the request is not valid in the current state.
          content:
            application/json:
              schema: *a1
        "401":
          description: No authenticated consumer, or admin credentials are required.
          content:
            application/json:
              schema: *a1
        "403":
          description: The API key lacks the required scope, or the request did not arrive
            through the gateway.
          content:
            application/json:
              schema: *a1
        "404":
          description: The resource does not exist, or does not belong to this consumer.
          content:
            application/json:
              schema: *a1
        "409":
          description: Conflicts with existing state — an idempotency clash, an operation
            already in flight, or an invalid state transition.
          content:
            application/json:
              schema: *a1
        "500":
          description: Unexpected server error. The detail is logged server-side and never
            returned.
          content:
            application/json:
              schema: *a1
        "502":
          description: An upstream provider returned an error.
          content:
            application/json:
              schema: *a1
        "503":
          description: An upstream provider is unavailable, or the service is misconfigured.
          content:
            application/json:
              schema: *a1
      summary: Delete a receiver
      tags: *a11
  /v1/kyc/receivers/{id}/approve:
    post:
      operationId: ReceiversController_approve_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveReceiverDto"
      responses:
        "200":
          description: Receiver approved; terms link returned
        "400":
          description: Validation failed, or the request is not valid in the current state.
          content:
            application/json:
              schema: *a1
        "401":
          description: No authenticated consumer, or admin credentials are required.
          content:
            application/json:
              schema: *a1
        "403":
          description: The API key is not elevated (admin role). A key may submit KYC data
            or approve it, not both.
        "404":
          description: The resource does not exist, or does not belong to this consumer.
          content:
            application/json:
              schema: *a1
        "409":
          description: Invalid KYC status transition (e.g. receiver is not in
            'pending_review')
        "500":
          description: Unexpected server error. The detail is logged server-side and never
            returned.
          content:
            application/json:
              schema: *a1
        "502":
          description: An upstream provider returned an error.
          content:
            application/json:
              schema: *a1
        "503":
          description: An upstream provider is unavailable, or the service is misconfigured.
          content:
            application/json:
              schema: *a1
      summary: Approve a pending-review receiver (admin-only review gate); sends the
        customer the terms link and returns it
      tags: *a11
  /v1/kyc/receivers/{id}/tos:
    post:
      operationId: ReceiversController_requestTos_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestTosDto"
      responses:
        "201":
          description: ""
        "400":
          description: Validation failed, or the request is not valid in the current state.
          content:
            application/json:
              schema: *a1
        "401":
          description: No authenticated consumer, or admin credentials are required.
          content:
            application/json:
              schema: *a1
        "403":
          description: The API key lacks the required scope, or the request did not arrive
            through the gateway.
          content:
            application/json:
              schema: *a1
        "404":
          description: The resource does not exist, or does not belong to this consumer.
          content:
            application/json:
              schema: *a1
        "409":
          description: Conflicts with existing state — an idempotency clash, an operation
            already in flight, or an invalid state transition.
          content:
            application/json:
              schema: *a1
        "500":
          description: Unexpected server error. The detail is logged server-side and never
            returned.
          content:
            application/json:
              schema: *a1
        "502":
          description: An upstream provider returned an error.
          content:
            application/json:
              schema: *a1
        "503":
          description: An upstream provider is unavailable, or the service is misconfigured.
          content:
            application/json:
              schema: *a1
      summary: Request a terms-of-service link for a receiver ('code' returns the URL;
        'email' sends it, max once/day)
      tags: *a11
  /v1/kyc/receivers/{id}/enable:
    post:
      operationId: ReceiversController_enable_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnableReceiverDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiverEntity"
        "400":
          description: Validation failed, or the request is not valid in the current state.
          content:
            application/json:
              schema: *a1
        "401":
          description: No authenticated consumer, or admin credentials are required.
          content:
            application/json:
              schema: *a1
        "403":
          description: The API key lacks the required scope, or the request did not arrive
            through the gateway.
          content:
            application/json:
              schema: *a1
        "404":
          description: The resource does not exist, or does not belong to this consumer.
          content:
            application/json:
              schema: *a1
        "409":
          description: Invalid KYC status transition (e.g. receiver is not in
            'pending_user')
        "500":
          description: Unexpected server error. The detail is logged server-side and never
            returned.
          content:
            application/json:
              schema: *a1
        "502":
          description: An upstream provider returned an error.
          content:
            application/json:
              schema: *a1
        "503":
          description: An upstream provider is unavailable, or the service is misconfigured.
          content:
            application/json:
              schema: *a1
      summary: Enable an inactive receiver with an accepted terms-of-service id
      tags: *a11
  /v1/kyc/receivers/{id}/access:
    patch:
      operationId: ReceiversController_setAccess_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetAccessDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiverEntity"
        "400":
          description: Validation failed, or the request is not valid in the current state.
          content:
            application/json:
              schema: *a1
        "401":
          description: No authenticated consumer, or admin credentials are required.
          content:
            application/json:
              schema: *a1
        "403":
          description: The API key is not elevated (admin role); a tenant key cannot lift
            a kill-switch an operator applied.
        "404":
          description: The resource does not exist, or does not belong to this consumer.
          content:
            application/json:
              schema: *a1
        "409":
          description: Conflicts with existing state — an idempotency clash, an operation
            already in flight, or an invalid state transition.
          content:
            application/json:
              schema: *a1
        "500":
          description: Unexpected server error. The detail is logged server-side and never
            returned.
          content:
            application/json:
              schema: *a1
        "502":
          description: An upstream provider returned an error.
          content:
            application/json:
              schema: *a1
        "503":
          description: An upstream provider is unavailable, or the service is misconfigured.
          content:
            application/json:
              schema: *a1
      summary: Enable or disable a fiat account (admin-only kill-switch for
        onramp/offramp)
      tags: *a11
...
info:
  title: Cosmos Pay — Payments API
  description: "Payments microservice (Stellar payment intents). All endpoints
    require traffic to arrive through the APISIX gateway: a valid
    `X-Gateway-Secret` header plus an authenticated consumer
    (`X-Consumer-Username`). Paths already include the version (`/v1/...`)."
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    gateway-secret:
      type: apiKey
      in: header
      name: X-Gateway-Secret
    consumer:
      type: apiKey
      in: header
      name: X-Consumer-Username
  schemas:
...
    CreateReceiverDto:
      type: object
      properties:
        type:
          type: string
          enum: &a22
            - individual
            - business
          example: individual
        kyc_type:
          type: string
          enum: &a23
            - light
            - standard
            - enhanced
          example: standard
        email:
          type: string
          example: jane@acme.com
        country:
          type: string
          example: US
          description: ISO 3166-1 alpha-2 country code
        first_name:
          type: string
          example: Jane
        last_name:
          type: string
          example: Doe
        date_of_birth:
          type: string
          example: 1985-04-12T00:00:00.000Z
          description: ISO 8601 datetime (BlindPay rejects date-only values).
        tax_id:
          type: string
          example: 123-45-6789
        occupation:
          type: string
          example: Engineer
        id_doc_country:
          type: string
          example: US
        id_doc_type:
          type: string
          example: PASSPORT
        id_doc_front_file:
          type: string
          description: file_url from POST /v1/kyc/upload
        id_doc_back_file:
          type: string
        selfie_file:
          type: string
        account_purpose:
          type: string
          example: receiving_payments
        account_purpose_other:
          type: string
        source_of_funds_doc_type:
          type: string
          example: employment
        source_of_funds_doc_file:
          type: string
        source_of_wealth:
          type: string
        sole_proprietor_doc_type:
          type: string
          example: bank_statement
          description: master_service_agreement | salary_slip | bank_statement
        proof_of_address_doc_type:
          type: string
          example: UTILITY_BILL
        proof_of_address_doc_file:
          type: string
        recipient_relationship:
          type: string
          example: self
        purpose_of_transactions:
          type: string
          example: business_transactions
        purpose_of_transactions_explanation:
          type: string
        legal_name:
          type: string
          example: Acme Inc.
        alternate_name:
          type: string
          example: Acme
        business_type:
          type: string
          example: cooperative
        business_industry:
          type: string
          example: "541511"
        business_description:
          type: string
        formation_date:
          type: string
          example: 2015-01-01T00:00:00.000Z
        estimated_annual_revenue:
          type: string
          example: 1000000_4999999
        publicly_traded:
          type: boolean
          example: false
        website:
          type: string
          example: https://acme.com
        incorporation_doc_file:
          type: string
        proof_of_ownership_doc_file:
          type: string
        owners:
          type: array
          items:
            $ref: "#/components/schemas/ReceiverOwnerDto"
        address_line_1:
          type: string
          example: 123 Main St
        address_line_2:
          type: string
        city:
          type: string
          example: New York
        state_province_region:
          type: string
          example: NY
        postal_code:
          type: string
          example: "10001"
        phone_number:
          type: string
          example: "+12025550123"
        ip_address:
          type: string
        external_id:
          type: string
          description: Your own id for this receiver
        image_url:
          type: string
        tos_id:
          type: string
          description: Accepted terms-of-service id (from the redirect of POST
            /v1/kyc/terms-of-service). Required by BlindPay to create a
            receiver.
      required:
        - type
        - kyc_type
        - email
        - country
    ReceiverEntity:
      type: object
      properties:
        id:
          type: string
          example: clz9xreceiver01
        blindpayId:
          type: string
          example: re_000000000000
        type:
          type: string
          example: individual
        kycType:
          type: object
          example: standard
        kycStatus:
          type: object
          example: verifying
          description: BlindPay KYC status (verifying, approved, rejected, ...).
        email:
          type: object
          example: jane@acme.com
        name:
          type: object
          example: Jane Doe
        country:
          type: object
          example: US
        externalId:
          type: object
          example: cust_001
        disabled:
          type: boolean
          example: false
          description: "Owner/admin kill-switch: when true the account is blocked from
            onramp/offramp."
        createdAt:
          format: date-time
          type: string
          example: 2026-06-28T12:00:00.000Z
        updatedAt:
          format: date-time
          type: string
          example: 2026-06-28T12:00:00.000Z
      required:
        - id
        - blindpayId
        - type
        - disabled
        - createdAt
        - updatedAt
    ReceiverListEntity:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ReceiverEntity"
        total:
          type: number
          example: 1
      required:
        - data
        - total
    ApproveReceiverDto:
      type: object
      properties:
        redirect_url:
          type: string
          example: https://dev.cosmospay.lat/kyc/return/org/dev/clz9xreceiver01
          description: Where BlindPay redirects the customer after they accept the terms.
            Must be https and on the consumer allow-list.
      required:
        - redirect_url
    RequestTosDto:
      type: object
      properties:
        channel:
          type: string
          enum:
            - code
            - email
          default: code
          description: "'code' returns the hosted ToS URL to display; 'email' sends it to
            the receiver (max once per day)."
        redirect_url:
          type: string
          example: https://yourapp.com/kyc/return
          description: Where BlindPay redirects after acceptance (gets ?tos_id=...). Must
            be https and on the consumer allow-list.
      required:
        - redirect_url
    EnableReceiverDto:
      type: object
      properties:
        tos_id:
          type: string
          example: to_a1b2c3d4e5f6g7h
          description: Accepted terms-of-service id (the ?tos_id=... from the ToS redirect).
      required:
        - tos_id
...
    UpdateReceiverDto:
      type: object
      properties:
        type:
          type: string
          enum: *a22
          example: individual
        kyc_type:
          type: string
          enum: *a23
          example: standard
        email:
          type: string
          example: jane@acme.com
        country:
          type: string
          example: US
          description: ISO 3166-1 alpha-2 country code
        first_name:
          type: string
          example: Jane
        last_name:
          type: string
          example: Doe
        date_of_birth:
          type: string
          example: 1985-04-12T00:00:00.000Z
          description: ISO 8601 datetime (BlindPay rejects date-only values).
        tax_id:
          type: string
          example: 123-45-6789
        occupation:
          type: string
          example: Engineer
        id_doc_country:
          type: string
          example: US
        id_doc_type:
          type: string
          example: PASSPORT
        id_doc_front_file:
          type: string
          description: file_url from POST /v1/kyc/upload
        id_doc_back_file:
          type: string
        selfie_file:
          type: string
        account_purpose:
          type: string
          example: receiving_payments
        account_purpose_other:
          type: string
        source_of_funds_doc_type:
          type: string
          example: employment
        source_of_funds_doc_file:
          type: string
        source_of_wealth:
          type: string
        sole_proprietor_doc_type:
          type: string
          example: bank_statement
          description: master_service_agreement | salary_slip | bank_statement
        proof_of_address_doc_type:
          type: string
          example: UTILITY_BILL
        proof_of_address_doc_file:
          type: string
        recipient_relationship:
          type: string
          example: self
        purpose_of_transactions:
          type: string
          example: business_transactions
        purpose_of_transactions_explanation:
          type: string
        legal_name:
          type: string
          example: Acme Inc.
        alternate_name:
          type: string
          example: Acme
        business_type:
          type: string
          example: cooperative
        business_industry:
          type: string
          example: "541511"
        business_description:
          type: string
        formation_date:
          type: string
          example: 2015-01-01T00:00:00.000Z
        estimated_annual_revenue:
          type: string
          example: 1000000_4999999
        publicly_traded:
          type: boolean
          example: false
        website:
          type: string
          example: https://acme.com
        incorporation_doc_file:
          type: string
        proof_of_ownership_doc_file:
          type: string
        owners:
          type: array
          items:
            $ref: "#/components/schemas/ReceiverOwnerDto"
        address_line_1:
          type: string
          example: 123 Main St
        address_line_2:
          type: string
        city:
          type: string
          example: New York
        state_province_region:
          type: string
          example: NY
        postal_code:
          type: string
          example: "10001"
        phone_number:
          type: string
          example: "+12025550123"
        ip_address:
          type: string
        external_id:
          type: string
          description: Your own id for this receiver
        image_url:
          type: string
        tos_id:
          type: string
          description: Accepted terms-of-service id (from the redirect of POST
            /v1/kyc/terms-of-service). Required by BlindPay to create a
            receiver.
...
security:
  - gateway-secret: []
  - consumer: []